refactor(audit): move worktree process execution into core service (#5879)
